Handle embedded structs on fieldMap reflection
Embedded structs are presented to reflect as `Anonymous` fields, the
transparency offered to the user in regular go syntax does not apply
here.

To ensure we manage to assign the correct SQL column to the right
type field upon Scan we do a recursive unwrap of the embedded structs
making their inner fields available to the scanner as if they were
part of the base type.